Update README.md
This repository is a my submission for Getting and Cleaning Data course project. It has the instructions on how to run analysis on Human Activity recognition dataset.
CodeBook.md is a code book that describes the variables, data, and work performed by me to clean up the data
run_analysis.R performs the data preparation (like library dplyr inclusion and etc) and then followed by the 5 steps required as described in the course project’s definition:
Step1. Merging the training and the test sets to create one data set.
Step2. Extracting only the measurements on the mean and standard deviation for each measurement.
Step3. Using descriptive activity names to name the activities in the data set
Step4. Appropriately labeling the data set with descriptive variable names.
Step5, From the data set in step 4, a second, independent tidy data set with the average of each variable for each activity and each subject was created.
finalData.txt is the exported final data after going through all the sequences described above.